requirements.txt
appdirs==1.4.3
args==0.1.0
clint==0.5.1
docopt==0.6.2
jsonpatch==0.4
packaging==16.8
pyparsing==2.2.0
PyYAML==3.12
requests==2.13.0
schema==0.5.0
six==1.10.0
